Factor out search-autocomplete.tsx.
