+++ package.json
@@ -74,7 +74,7 @@
     "@babel/preset-env": "7.7.6",
     "@babel/preset-typescript": "7.7.4",
     "@babel/register": "7.7.4",
-    "@justfixnyc/geosearch-requester": "0.0.4",
+    "@justfixnyc/geosearch-requester": "0.1.0",
     "@justfixnyc/react-aria-modal": "5.1.0",
     "@justfixnyc/util": "^0.0.1",
     "@loadable/babel-plugin": "5.10.0",
